server/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
pydantic==2.12.5
google-genai==1.70.0
python-multipart==0.0.24
pypdf==6.9.2
aiofiles==25.1.0
